==> project_seodam/preprocessing/1.Crawl_seodam.ipynb <==
import time
from collections import OrderedDict
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support import expected_conditions as EC

### Crawl comments and postings

chromedriver = './private/files/chromedriver'
driver = webdriver.Chrome(chromedriver)
driver.get('http://www.ssodam.com/')
wait = WebDriverWait(driver,10)

try:
    button = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "button[data-dismiss='modal']")))
    button.click()
except:
    pass

button2 = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "a[data-target='#loginModal']")))
time.sleep(1)
button2.click()

email = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "input[class='form-control']")))
email.clear()
email.send_keys('****')


password = driver.find_element_by_css_selector("input[placeholder='Password']")
password.clear()
password.send_keys('****')

login = driver.find_element_by_css_selector("input[class='submit btn btn-default']")
login.click()

time.sleep(1)

alert = driver.switch_to_alert()
alert.accept()

crawl = OrderedDict()
crawl['title'] = []
crawl['body'] = []
crawl['date'] = []

comments_dict = OrderedDict()
comments_dict['text'] = []

start=238249; end=243589
for i in range(start, end):
    
    if i%1000==0:
        print '{}/243589 done'.format(i)
        
    driver.get('http://www.ssodam.com/content/{}'.format(i))
    try:
        title = driver.find_element_by_xpath("//div[@class='board-title']")
        body = driver.find_element_by_xpath("//div[@class='board-content mverdana']")
        date = driver.find_element_by_xpath("//div[@class='board-date row col-lg-3 col-xs-12 desktop-hide']")
    except:
        continue
    
    try:
        crawl['title'].append(title.text.encode('utf-8'))
        crawl['body'].append(body.text.encode('utf-8'))
        crawl['date'].append(date.text.encode('utf-8'))
    except :
        continue

    try:
        comment_num = driver.find_element_by_xpath("//div[@class='comment-number']")
    except:
        continue
        
    if comment_num.text == u'댓글 0개':
        continue

    div = driver.find_element_by_xpath("//div[@id='comments_area_html']")
    div2 = div.find_elements_by_xpath("//div[@class='mverdana comment-content']")
    div3 = div.find_elements_by_xpath("//div[@class='comment-date']")
    div4 = div.find_elements_by_xpath("//div[@class='mverdana comment-content comment-user-writer']")

    for comments in div2 :
        comments_dict['text'].append(comments.text.encode('utf-8'))
    
    for comments in div4 :
        comments_dict['text'].append(comments.text.encode('utf-8'))
    
df_crawl = pd.DataFrame(crawl)
df_comments = pd.DataFrame(comments_dict)
df_crawl.to_csv('crawl{}_{}.csv'.format(start,end))
df_comments.to_csv('comments{}_{}.csv'.format(start,end))

### Crawl comments only

chromedriver = './private/files/chromedriver'
driver = webdriver.Chrome(chromedriver)
driver.get('http://www.ssodam.com/')
wait = WebDriverWait(driver,10)

try:
    button = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "button[data-dismiss='modal']")))
    button.click()
except:
    pass

button2 = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "a[data-target='#loginModal']")))
time.sleep(1)
button2.click()

email = wait.until(EC.visibility_of_element_located((By.CSS_SELECTOR, "input[class='form-control']")))
email.clear()
email.send_keys('****')

password = driver.find_element_by_css_selector("input[placeholder='Password']")
password.clear()
password.send_keys('****')

login = driver.find_element_by_css_selector("input[class='submit btn btn-default']")
login.click()

time.sleep(1)

alert = driver.switch_to_alert()
alert.accept()

start = 200294; end = 210294
comments_dict = OrderedDict()
comments_dict['text'] = []
#comments_dict['date'] = []

for i in range(start, end):
    driver.get('http://www.ssodam.com/content/{}'.format(i))
    try:
        comment_num = driver.find_element_by_xpath("//div[@class='comment-number']")
    except:
        continue
    
    if comment_num.text == u'댓글 0개':
        continue

    div = driver.find_element_by_xpath("//div[@id='comments_area_html']")
    div2 = div.find_elements_by_xpath("//div[@class='mverdana comment-content']")
    div3 = div.find_elements_by_xpath("//div[@class='comment-date']")
    div4 = div.find_elements_by_xpath("//div[@class='mverdana comment-content comment-user-writer']")

    for comments in div2 :
        comments_dict['text'].append(comments.text.encode('utf-8'))
    
    for comments in div4 :
        comments_dict['text'].append(comments.text.encode('utf-8'))
    
'''    for date0 in div3 :
        if not date0.get_attribute('style') :
            date = '2017' + '-' + re.sub(r'/', '-', date0.text)
            comments_dict['date'].append(date)
'''
df = pd.DataFrame(comments_dict)
df.to_csv('./private/files/reply/crawl_reply{}_{}.csv'.format(start,end))
