==> project_seodam/preprocessing/2.Crawled_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np
from konlpy.utils import pprint
from bs4 import BeautifulSoup
import re

df = pd.read_csv('../private/files/main/crawl238249_243589.csv').drop('Unnamed: 0', axis=1)

### 게시글 preprocessing

#df_crawl : DataFrame with columns 'title', 'body'
def preprocessing(df_crawl):
    df_frozen=pd.read_csv('../private/files/seodam_cleaned_short0307.csv')

    df_crawl.reset_index(drop=True, inplace=True)
    
    #remove empty rows
    df_crawl = df_crawl.ix[df_crawl['title']!='empty']
    
    #delete 'frozen' postings
    frozen_list = df_frozen['제목']
    df_crawl = df_crawl.ix[~df_crawl['title'].isin(frozen_list)]
    
    #merge cells
    df_crawl['text'] = df_crawl['title'].map(str) + ' ' + df_crawl['body']
    df_crawl.drop(['title', 'body'], axis=1, inplace=True)
    df_crawl.reset_index(drop=True, inplace=True)
    
    #remove html
    temp = []
    for i in range(len(df_crawl['text'])):
        try:
            remove = re.sub(r'[\n\r]', ' ', df_crawl.ix[i,'text'])
            remove = remove
            temp.append(remove)
        except:
            temp.append('empty')
    df_crawl['text2'] = temp
    df_crawl.drop('text', axis=1, inplace=True)
    
    #remove empty rows
    df_crawl = df_crawl.ix[df_crawl['text2']!='empty']
    df_crawl.reset_index(drop=True, inplace=True)
    
    return df_crawl

result = preprocessing(df)

result.to_csv('../private/files/main/unfrozen238249_243589.csv')

### 댓글 preprocessing

df4 = pd.read_csv('../private/files/comments/crawl_reply193294_194294.csv').drop('Unnamed: 0', axis=1)
df5 = pd.read_csv('../private/files/comments/crawl_reply194294_195294.csv').drop('Unnamed: 0', axis=1)
df6 = pd.read_csv('../private/files/comments/crawl_reply195294_196294.csv').drop('Unnamed: 0', axis=1)
df7 = pd.read_csv('../private/files/comments/crawl_reply196294_197294.csv').drop('Unnamed: 0', axis=1)
df8 = pd.read_csv('../private/files/comments/crawl_reply198294_199294.csv').drop('Unnamed: 0', axis=1)
df9 = pd.read_csv('../private/files/comments/crawl_reply199294_200294.csv').drop('Unnamed: 0', axis=1)
df10 = pd.read_csv('../private/files/comments/crawl_reply200294_210294.csv').drop('Unnamed: 0', axis=1)
df1 = pd.read_csv('../private/files/comments/comments210294_223249.csv').drop('Unnamed: 0', axis=1)
df2 = pd.read_csv('../private/files/comments/comments223249_238249.csv').drop('Unnamed: 0', axis=1)
df3 = pd.read_csv('../private/files/comments/comments238249_243589.csv').drop('Unnamed: 0', axis=1)

df = pd.concat([df4, df5, df6, df7, df8, df9, df10, df1, df2, df3], axis=0)

def preprocessing_comments(df):
    df_frozen = pd.read_csv('../private/files/seodam_re_notags0323.csv').drop('Unnamed: 0', axis=1)
    
    df.reset_index(drop=True, inplace=True)
    
    #remove empty rows
    lst = []
    for i, text in enumerate(df['text']):
        if type(text) == float:
            lst.append(i)
    df.drop(lst, axis=0, inplace=True)
    df.reset_index(drop=True, inplace=True)
    
    #remove revmoved rows
    df = df.ix[df['text']!='삭제된 댓글입니다.']
    df.reset_index(drop=True, inplace=True)
    
    #remove tags
    temp = []
    for i in range(len(df['text'])):
        try:
            remove = re.sub(r'[\n\r]', ' ', df.ix[i,'text'])
            remove = remove
            temp.append(remove)
        except:
            temp.append('empty')
    df['text2'] = temp
    df.drop('text', axis=1, inplace=True)
    
    #remove frozen comments
    frozen_list = df_frozen['text']
    df = df.ix[~df['text2'].isin(frozen_list)]
    df.reset_index(drop=True, inplace=True)
    
    return df

result = preprocessing_comments(df)

result.to_csv('../private/files/comments/unfrozen_comments0403.csv')
