==> project_seodam/preprocessing/3.Frozen_data_preprocessing.ipynb <==
from bs4 import BeautifulSoup

### bs4를 통해 게시글에서 태그 제거

df = pd.read_csv('./files/seodam_notnull.csv')

def remove_tags(df):
    new_col = []
    for i in range(1316):
        try:
            soup = BeautifulSoup(df.ix[i,'내용'], from_encoding = 'utf-8')
            text = soup.get_text().strip()
            new_col.append(text)
        except:
            new_col.append('error')
    return new_col

df2 = pd.DataFrame(remove_tags(df), columns=['body'])

writer = pd.ExcelWriter('remove_tag(test).xlsx', engine=None)
df2.to_excel(writer, sheet_name='1', encoding ='utf-8')      # 결과를 엑셀로 저장
writer.save()

### 빈 게시글 제거

df_del = pd.read_csv('./files/seodam_del.csv')

df_del = df_del.drop([48])
to_del = [467]
for i in range(56):    
    x = np.array(df_del)[i,0]
    to_del.append(x)

df_droped = df.drop(to_del)        #태그 제거 안된 data
df2_droped = df2.drop(to_del)      #태그 제거된 data

### clean_data_seodam.csv에 태그 제거된 data 저장 (제목, 내용, 냉동 개수, 날짜)

clean_data_seodam = pd.concat([df_droped, df2_droped], axis = 1)
clean_data_seodam2 = clean_data_seodam.drop(['내용',0], axis = 1)

writer = pd.ExcelWriter('seodam_cleaned0307.xlsx', engine=None)
clean_data_seodam.to_excel(writer, sheet_name='1', encoding ='utf-8')      # 결과를 엑셀로 저장
writer.save()

writer = pd.ExcelWriter('seodam_cleaned_short0307.xlsx', engine=None)
clean_data_seodam2.to_excel(writer, sheet_name='1', encoding ='utf-8')      # 결과를 엑셀로 저장
writer.save()

### 제목과 내용을 한 column으로 합친다

df33 = pd.read_csv('seodam_cleaned_short0307.csv')
df33 = df33.drop(['Unnamed: 0'], axis=1)

df33['text'] = df33['제목'].map(str) + ' ' + df33['body']

seodam_preprocess_done0307 = df33.drop(['제목', 'body'], axis = 1)

seodam_preprocess_done0307.to_csv('seodam_preprocess_done0307.csv')

### 댓글 태그 제거

reply0 = pd.read_csv('./files/seodam_re_eidited.csv')

reply0.columns = ['text', 'freeze']

df1 = reply0['text']
df2 = reply0['freeze']
no_tags = []
for i in df1:
    soup = BeautifulSoup(i, from_encoding='utf-8')
    no_tags.append(soup.get_text().encode('utf-8'))
df3 = pd.DataFrame(no_tags, columns=['text'])
reply = pd.concat([df2, df3], axis = 1)

reply = reply.drop(reply.index[50])

reply.to_csv('./files/seodam_re_notags0323.csv')
