==> project_seodam/preprocessing/4.Concat_frozen_dataset.ipynb <==
reply = pd.read_csv('../private/files/seodam_re_notags0323.csv').ix[:,1:]
main = pd.read_csv('../private/files/seodam_preprocess_done0307.csv').ix[:,1:].drop(['날짜'], axis=1)

main.columns = ['freeze', 'text']

lst = []
for text in main['text']:
    temp = re.sub(r'[\r\n]', ' ', text)
    lst.append(temp)
df = pd.DataFrame(lst, columns=['text'])

main2 = pd.concat([main['freeze'], df], axis=1)
put_together = pd.concat([main2, reply], axis=0)
put_together.reset_index(drop=True, inplace=True)

put_together.set_value([450, 1109], 'freeze', [2,5])
put_together.drop([451, 1110], axis=0, inplace=True)
put_together.reset_index(drop=True, inplace=True)

put_together.to_csv('./private/files/seodam_together_notags0326.csv')
