==> project_seodam/preprocessing/5.Make_random_unfrozen_samples.ipynb <==
main0 = pd.read_csv('../private/files/main/unfrozen0402.csv').drop(['Unnamed: 0', 'Unnamed: 0.1'], axis=1)
comments0 = pd.read_csv('../private/files/comments/unfrozen_comments0403.csv').drop(['Unnamed: 0'], axis=1)

main= main0.sample(1259, random_state=0)
comments = comments0.sample(1954, random_state=0)

unfrozen= pd.concat([main, comments], axis=0)

unfrozen.drop('date', axis=1, inplace=True)

unfrozen.reset_index(drop=True, inplace=True)

unfrozen.to_csv('../private/files/unfrozen_mixed0402')
